@@ -6,17 +6,17 @@
     "preview": "pnpm build && vite preview",
     "lint": "pnpm check & npx eslint .",
     "check": "svelte-kit sync && npx svelte-check --tsconfig ./tsconfig.json",
-    "format": "npx @jill64/psx"
+    "format": "npx psvx"
   },
   "devDependencies": {
-    "@jill64/eslint-config-svelte": "1.3.6",
+    "@jill64/eslint-config-svelte": "2.0.2",
     "@fontsource/fira-mono": "5.1.0",
     "@jill64/sentry-sveltekit-edge": "1.2.55",
     "@neoconfetti/svelte": "2.2.1",
-    "@sveltejs/kit": "2.13.0",
+    "@sveltejs/kit": "2.15.1",
     "@types/cookie": "1.0.0",
-    "svelte": "4.2.19",
-    "vite": "5.4.11",
-    "@sveltejs/vite-plugin-svelte": "3.1.2"
+    "svelte": "5.16.0",
+    "vite": "6.0.6",
+    "@sveltejs/vite-plugin-svelte": "5.0.3"
   }
 }
