@@ -1,11 +1,15 @@
 <script lang="ts">
-  import { spring } from 'svelte/motion'
+  import { Spring } from 'svelte/motion'
 
-  let count = 0
+  let count = $state(0)
 
-  const displayed_count = spring()
-  $: displayed_count.set(count)
-  $: offset = modulo($displayed_count, 1)
+  let displayed_count = new Spring(0)
+
+  $effect(() => {
+    displayed_count.set(count)
+  })
+
+  let offset = $derived(modulo(displayed_count.current, 1))
 
   function modulo(n: number, m: number) {
     // handle negative numbers
@@ -14,10 +18,7 @@
 </script>
 
 <div class="counter">
-  <button
-    on:click={() => (count -= 1)}
-    aria-label="Decrease the counter by one"
-  >
+  <button onclick={() => (count -= 1)} aria-label="Decrease the counter by one">
     <svg aria-hidden="true" viewBox="0 0 1 1">
       <path d="M0,0.5 L1,0.5" />
     </svg>
@@ -29,16 +30,13 @@
       style="transform: translate(0, {100 * offset}%)"
     >
       <strong class="hidden" aria-hidden="true"
-        >{Math.floor($displayed_count + 1)}</strong
+        >{Math.floor(displayed_count.current + 1)}</strong
       >
-      <strong>{Math.floor($displayed_count)}</strong>
+      <strong>{Math.floor(displayed_count.current)}</strong>
     </div>
   </div>
 
-  <button
-    on:click={() => (count += 1)}
-    aria-label="Increase the counter by one"
-  >
+  <button onclick={() => (count += 1)} aria-label="Increase the counter by one">
     <svg aria-hidden="true" viewBox="0 0 1 1">
       <path d="M0,0.5 L1,0.5 M0.5,0 L0.5,1" />
     </svg>
