@@ -1,4 +1,7 @@
 /* $$__ADAPTER_IMPORTS__$$ */
 import * as cdk from 'aws-cdk-lib'
 
-export const lambdaModifier = (lambdaFunction: cdk.aws_lambda.Function) => {} /* $$__LAMBDA_MODIFIER__$$ */
+export const lambdaModifier = (
+  // eslint-disable-next-line
+  lambdaFunction: cdk.aws_lambda.Function
+) => {} /* $$__LAMBDA_MODIFIER__$$ */
