@@ -71,7 +71,7 @@ export const setup = async ({ builder, tmp, options }: Context) => {
       __DOMAIN_NAME__: options.domain?.fqdn ?? '',
       __CERTIFICATE_ARN__: options.domain?.certificateArn ?? '',
       __LAMBDA_RUNTIME__: options.runtime ?? 'NODE_LATEST',
-      '{} /* $$__ENVIRONMENT__$$ */': JSON.stringify(options.env ?? {}),
+      '{} /* $$__ENVIRONMENT__$$ */': JSON.stringify(options.env ?? {})
     }
   )
 
@@ -80,7 +80,9 @@ export const setup = async ({ builder, tmp, options }: Context) => {
     path.join(options.out, 'external', 'cdk-modifiers.ts'),
     {
       '/* $$__ADAPTER_IMPORTS__$$ */': options.adapterImports?.join('\n') ?? '',
-      '(lambdaFunction: cdk.aws_lambda.Function) => {} /* $$__LAMBDA_MODIFIER__$$ */': options.lambdaModifier?.toString() ?? '(lambdaFunction: cdk.aws_lambda.Function) => {}'
+      '(lambdaFunction: cdk.aws_lambda.Function) => {} /* $$__LAMBDA_MODIFIER__$$ */':
+        options.lambdaModifier?.toString() ??
+        '(lambdaFunction: cdk.aws_lambda.Function) => {}'
     }
   )
 
